requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.2
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
autocommand==2.2.2
autopep8==2.0.4
bcrypt==4.0.1
beautifulsoup4==4.10.0
billiard==4.1.0
black==23.3.0
Brotli==1.0.9
celery==5.3.4
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.10
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
configparser==6.0.0
distlib==0.3.7
Django==4.2.7
django-debug-toolbar==4.0.0
django-filter==23.1
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
exceptiongroup==1.1.3
filelock==3.12.2
flake8==6.0.0
Flask==2.0.2
frozenlist==1.3.3
future==0.18.2
gevent==23.9.1
gprof2dot==2022.7.29
greenlet==2.0.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
inflect==7.0.0
iniconfig==2.0.0
installer==0.4.0
irc==20.3.0
itsdangerous==2.0.1
jaraco.collections==4.3.0
jaraco.context==4.3.0
jaraco.functools==4.0.0
jaraco.logging==3.2.0
jaraco.stream==3.0.3
jaraco.text==3.11.1
Jinja2==3.0.3
kombu==5.3.2
konfig==1.1
loads==0.3
MarkupSafe==2.0.1
mccabe==0.7.0
more-itertools==9.1.0
MouseInfo==0.1.3
multidict==6.0.4
MyApplication==0.1.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
pefile==2021.9.3
phone==0.4.3
Pillow==9.0.0
platformdirs==3.9.1
pluggy==1.3.0
prompt-toolkit==3.0.39
psutil==5.9.6
psycopg2-binary==2.9.5
py-youtube==1.1.7
PyAutoGUI==0.9.53
pycares==4.3.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.9
pyflakes==3.0.1
pygame==2.1.2
PyGetWindow==0.0.9
pyinstall==0.1.4
pyinstaller==4.8
pyinstaller-hooks-contrib==2021.5
PyMsgBox==1.0.7
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.26
pyTelegramBotAPI==4.3.1
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.4
pytz==2018.5
pywhatkit==5.3
pywin32-ctypes==0.2.0
PyYAML==6.0.1
pyzmq==25.1.1
redis==5.0.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.47
sqlparse==0.4.2
tempora==5.5.0
tomli==2.0.1
typing_extensions==4.6.3
tzdata==2023.3
ujson==5.8.0
urllib3==1.26.8
vine==5.0.0
virtualenv==20.24.3
waitress==2.1.2
wcwidth==0.2.9
webencodings==0.5.1
WebOb==1.8.7
WebTest==3.0.0
Werkzeug==2.0.2
wikipedia==1.4.0
ws4py==0.5.1
WSGIProxy2==0.5.1
yarl==1.9.2
zope.event==5.0
zope.interface==6.1
